+++ src/functions/selectChange.js
@@ -1,4 +1,5 @@
 import drawing from './../drawing.js';
+import colors from './../colors.js';
 
 window.addEventListener("load", function () {
   var selects = document.getElementsByTagName("select");
@@ -11,6 +12,10 @@ function selectHandler(select) {
 
 export default function selectChange(e) {
   var youClicked = {
+    "color-theme": function () {
+      colors.theme(e.target.value);
+      drawing.refresh();
+    },
     "grid": function () {
       drawing.drawing.grid = 96 / e.target.value;
       drawing.drawing.snap = drawing.drawing.grid / drawing.drawing.divisions * drawing.drawing.gridsnap;
@@ -47,11 +52,11 @@ export default function selectChange(e) {
       drawing.drawing.endsnap = e.target.value;
       drawing.refresh();
     },
-    "line_colour": function() {
-
+    "line_color": function() {
+      // this is just here to not throw an error for undefined function
     },
     "line_width": function() {
-
+      // ditto
     }
   }
   youClicked[e.target.id]();